+++ simple-calculator/Operators.cs
@@ -20,7 +20,7 @@ namespace simple_calculator
         {
             return x * y;
         }
-        public int divide(int x, int y)
+        public double divide(double x, double y)
         {
             return x / y;
         }
